+++ source/engine/controllers/Controller.cpp
@@ -51,7 +51,7 @@ void Controller::Move(float deltaTime)
 {
 	if (!_complete_path)
 	{
-		float distance = std::sqrtf(std::powf(_goal.x - _controlled_pawn->GetPosition().x, 2) + std::powf(_goal.y - _controlled_pawn->GetPosition().y, 2));
+		float distance = std::sqrt(std::pow(_goal.x - _controlled_pawn->GetPosition().x, 2) + std::pow(_goal.y - _controlled_pawn->GetPosition().y, 2));
 		_move_vector = (_goal - _controlled_pawn->GetPosition()) / distance;
 		if (_move_vector != glm::vec2(0.f, 0.f))
 		{
